refactor/ - added a context for managing filters - - the ContexForFilter module is located in the contex folder - - added FiltersProvider for state management - - removed the reducer function from the filters module and placed it in a new context -
