+++ src/App.tsx
@@ -1,21 +1,24 @@
-import { useState, useContext } from "react";
+import { useState, useContext, useReducer } from "react";
 import "./App.scss";
 import RegistrationModal from "./components/Registration/RegistrationModal";
 import AuthorizationModal from "./components/Authorization/Authorization";
 import Header from "./components/Header/Header";
 import Filters from "./components/Filters/Filters";
 import { ContextToken } from "./context/ContextForToken/ContextToken";
-
+import { FiltersProvider } from "./context/ContextForFlters/ContextForFilters";
 const App: React.FC = () => {
   const [isVisibleReg, setIsVisibleReg] = useState<boolean>(false);
   const [isVisibleAuth, setIsVisibleAuth] = useState<boolean>(false);
+
   const token = useContext(ContextToken);
   return (
     <>
       <ContextToken.Provider value={token}>
         {isVisibleReg && <RegistrationModal />}
         <Header />
-        <Filters />
+        <FiltersProvider>
+          <Filters />
+        </FiltersProvider>
       </ContextToken.Provider>
     </>
   );
