+++ src/components/Filters/Filters.tsx
@@ -1,32 +1,19 @@
-import { useState } from "react";
+import { useState, useContext } from "react";
 import Pagination from "../Pagination/Pagination";
 import imageCloseButton from "../../assets/free-icon-close-button.png";
 import "./Filters.scss";
 import MovieList from "../MovieList/MovieList";
 import Select from "./Select";
-import { optionAll } from "../../data/selectOptions/selectOptions";
 import { ChangeEvent } from "react";
-import { useReducer } from "react";
-
-interface Action {
-  type: "change";
-  event: string | number;
-}
-
-interface Option {
-  value: string | number;
-  label: string;
-}
+import { ContextOptionGenres } from "../../context/ContextForFlters/ContextForFilters";
+import { ContextOptionsYears } from "../../context/ContextForFlters/ContextForFilters";
+import { useChangeDispatch } from "../../context/ContextForFlters/ContextForFilters";
 
 const Filters = () => {
   const [selectedGenres, setSelectedGenres] = useState<Set<string>>(new Set());
-  const [optionsSelect, dispatch] = useReducer(yourReducer, optionAll);
-
-  const optionsGenres = optionAll.slice(0, 4);
-
-  const optionsYears = optionAll.slice(4, 8);
-
-  const curentValueGenres = optionsGenres.slice(3, 4);
+  const optionsGenres = useContext(ContextOptionGenres);
+  const optionsYears = useContext(ContextOptionsYears);
+  const dispatch = useChangeDispatch();
 
   const handleChange = (e: ChangeEvent<HTMLSelectElement>) => {
     dispatch({
@@ -85,17 +72,3 @@ const Filters = () => {
 };
 
 export default Filters;
-
-const yourReducer = (optionAll: Option[], action: Action) => {
-  switch (action.type) {
-    case "change": {
-      const selectedValue = action.event;
-      const selectedOption = optionAll.find(
-        (option) => option.value === selectedValue
-      );
-      return selectedOption ? [selectedOption] : optionAll;
-    }
-    default:
-      return optionAll;
-  }
-};
